Fix bug that prevented plonectl from starting in a non-zeo configuration.
Bug report from faro on IRC. After installing Plone 4.3b1 on Ubuntu (s)he got a
file not found error in subprocess upon running plonectl start. ctl.main
expects a server parameter that points to the ZEO executable, which it executes
on startup. If no zeo server should be started this is set to evaluate to
False, i.e. an empty string or a None.

However, when generating the plonectl script string interpolation is used, so
to ensure the path is interpreted as a string it is given as '%s', this causes
Nones to be coerced into 'None', which evaluates as True, so ctl attempts to
execute subprocess.popen("None"). Instead, I've changed it to include the repr()
of the path, which will be either '/usr/local/foobarbaz/zeoserver' or None.